feat: configure entire proyect to be compatible with common rules, common export and sass improvement
