+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev -p 7000",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
@@ -17,6 +17,8 @@
     "next": "13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "typescript": "4.9.5"
+    "sass": "^1.59.1",
+    "typescript": "4.9.5",
+    "yarn": "^1.22.19"
   }
 }
