+++ src/components/Header.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function Header() {
+  return (
+    <>
+      <h1>Nav Bar!!</h1>
+    </>
+  );
+}
