+++ src/pages/_app.page.tsx
@@ -1,0 +1,12 @@
+import "../styles/globals.scss";
+import type { AppProps } from "next/app";
+
+import { Layout } from "@common";
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
+}
