+++ src/pages/home/Home.module.scss
@@ -1,0 +1,5 @@
+@import "../../styles/settings/_settings";
+
+.title {
+  color: $color-primary;
+}
