+++ src/pages/home/Home.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import styles from "./Home.module.scss";
+
+export function HomePage() {
+  return (
+    <>
+      <h1 className={styles.title}>Home Page E</h1>
+    </>
+  );
+}
