+++ src/pages/index.page.tsx
@@ -1,0 +1,1 @@
+export { HomePage as default } from "./home/Home";
