+++ src/styles/settings/__settings.scss
@@ -1,0 +1,1 @@
+@import "colors";
