+++ src/styles/settings/_colors.scss
@@ -1,0 +1,1 @@
+$color-primary: blueviolet;
